requirements.txt
langchain_community 
langchain 
langgraph  
tavily-python
python-dotenv
langchain-google-genai
pydantic
langchain-experimental
fastapi
uvicorn
pandas
ta
yfinance
jinja2
python-multipart